renamed object folder to attributes and moved all items to base
