+++ OMF/OMF.csproj
@@ -64,30 +64,30 @@
     <Compile Include="base\ProjectElement.cs" />
     <Compile Include="base\Project.cs" />
     <Compile Include="base\Version.cs" />
-    <Compile Include="Objects\InvalidOmfException.cs" />
-    <Compile Include="Objects\Helpers.cs" />
-    <Compile Include="Objects\AttributeDefinitionArray.cs" />
-    <Compile Include="Objects\AttributeArray.cs" />
+    <Compile Include="base\InvalidOmfException.cs" />
+    <Compile Include="base\Helpers.cs" />
+    <Compile Include="attribute\AttributeDefinitionArray.cs" />
+    <Compile Include="attribute\AttributeArray.cs" />
     <Compile Include="volume\VolumeGridGeometry.cs" />
     <Compile Include="file\OMFWriter.cs" />
     <Compile Include="file\OMFReader.cs" />
     <Compile Include="base\ContentModel.cs" />
     <Compile Include="base\UidModel.cs" />
-    <Compile Include="Objects\Legend.cs" />
-    <Compile Include="Objects\ObjectFactory.cs" />
+    <Compile Include="data\Legend.cs" />
+    <Compile Include="base\ObjectFactory.cs" />
     <Compile Include="base\IObject.cs" />
-    <Compile Include="Objects\Int3Array.cs" />
+    <Compile Include="data\Int3Array.cs" />
     <Compile Include="lineset\LineSetGeometry.cs" />
-    <Compile Include="Objects\ScalarData.cs" />
-    <Compile Include="Objects\ScalarArray.cs" />
-    <Compile Include="Objects\ColorArray.cs" />
-    <Compile Include="Objects\StringArray.cs" />
+    <Compile Include="data\ScalarData.cs" />
+    <Compile Include="data\ScalarArray.cs" />
+    <Compile Include="data\ColorArray.cs" />
+    <Compile Include="data\StringArray.cs" />
     <Compile Include="lineset\LineSetElement.cs" />
-    <Compile Include="Objects\Int2Array.cs" />
+    <Compile Include="data\Int2Array.cs" />
     <Compile Include="volume\VolumeElement.cs" />
     <Compile Include="surface\SurfaceElement.cs" />
     <Compile Include="surface\SurfaceGeometry.cs" />
-    <Compile Include="Objects\Vector3Array.cs" />
+    <Compile Include="data\Vector3Array.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
