+++ OMF/base/InvalidOmfException.cs
