+++ OMF/data/ColorArray.cs
@@ -1,11 +1,8 @@
-﻿using System;
+﻿using OMF.Base;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-namespace OMF.Base
+
+namespace OMF
 {
     public class ColorArray : UidModel, IObject
     {
