+++ OMF/data/Int2Array.cs
@@ -1,27 +1,26 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using OMF.Base;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-namespace OMF.Base
+
+namespace OMF
 {
     public class Int2Array : UidModel, IObject
     {
+        [JsonIgnore]
+        public List<int[]> Data { get; set; }
+        public ScalarArray array { get; set; }
+
         public Int2Array()
         {
 
         }
+
         public Int2Array(List<int[]> data) : this()
         {
             Data = data;
         }
-        public ScalarArray array { get; set; }
-
-        [JsonIgnore]
-        public List<int[]> Data { get; set; }
-
+        
         public void Deserialize(Dictionary<string, object> json, System.IO.BinaryReader br)
         {
             Data = Helpers.ReadInt2Array(br, array);
