+++ OMF/data/ScalarArray.cs
@@ -1,11 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
-namespace OMF.Base
+using OMF.Base;
+
+namespace OMF
 {
     /// <summary>
     /// Class with unique ID and data array
