+++ OMF/data/ScalarData.cs
@@ -1,12 +1,8 @@
-﻿using System;
+﻿using OMF.Base;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Newtonsoft.Json;
 using System.IO;
 
-namespace OMF.Base
+namespace OMF
 {
     public class ScalarData : UidModel, IObject
     {
